Move converter and referencer into subfolders
- Move Converter and related modules into a `converter` subfolder
- Move Referencer and related modules into a `referencer` subfolder
