+++ lib/parser/converter/base-visitor.js
@@ -3,7 +3,7 @@
 const { replace, VisitorOption } = require("estraverse");
 const { matches, parse: parse_query } = require("esquery");
 
-const { KEYS } = require("./keys");
+const { KEYS } = require("../keys");
 
 class BaseVisitor {
   static *get_handlers(visitor) {
