+++ lib/parser/converter/converter.js
@@ -4,7 +4,7 @@ const { parse } = require("espree");
 const { getLineInfo } = require("acorn");
 const { VisitorOption } = require("estraverse");
 
-const { EXPRESSION_NODES } = require("./keys");
+const { EXPRESSION_NODES } = require("../keys");
 const { TAG_LOCATIONS } = require("./tag-locations");
 const { BaseVisitor } = require("./base-visitor");
 
@@ -314,14 +314,4 @@ class Converter extends BaseVisitor {
   }
 }
 
-function convert(code, options, ast) {
-  const ast_clone = JSON.parse(JSON.stringify(ast));
-  ast_clone.type = "TemplateRoot";
-
-  new Converter(code, options).visit(ast_clone);
-
-  // re-normalize to plain objects
-  return JSON.parse(JSON.stringify(ast_clone));
-}
-
-module.exports = { convert };
+module.exports = { Converter };
