+++ lib/parser/converter/tag-locations.js
