+++ lib/parser/referencer/referencer.js
@@ -1,8 +1,6 @@
 /* eslint-disable new-cap */
 "use strict";
-const { ScopeManager } = require("eslint-scope");
 const Referencer = require("eslint-scope/lib/referencer");
-const { KEYS } = require("./keys");
 const {
   AwaitScope,
   AwaitThenScope,
@@ -64,35 +62,4 @@ class SvelteReferencer extends Referencer {
   }
 }
 
-// https://github.com/eslint/eslint-scope/blob/14c092a6efd4dd0bf701bf4f8f518eac6b29b2ce/lib/index.js#L61-L76
-function defaultOptions() {
-  return {
-    optimistic: false,
-    directive: false,
-    nodejsScope: false,
-    impliedStrict: false,
-    sourceType: "script", // one of ['script', 'module']
-    ecmaVersion: 5,
-    childVisitorKeys: null,
-    fallback: "iteration"
-  };
-}
-
-function analyze(ast, providedOptions) {
-  const options = {
-    ...defaultOptions(),
-    ...providedOptions,
-    childVisitorKeys: KEYS
-  };
-  const manager = new ScopeManager(options);
-  const referencer = new SvelteReferencer(options, manager);
-  referencer.visit(ast);
-
-  if (manager.__currentScope !== null) {
-    throw new Error("currentScope should be null");
-  }
-
-  return manager;
-}
-
-module.exports = { analyze };
+module.exports = { SvelteReferencer };
