+++ lib/parser/referencer/scope.js
